Last hockey bot system update / changes
git-svn-id: http://www.twcore.org/svn/trunk@3672 9f8b3893-2a26-0410-9d00-a71e40fd2f77

+++ twcore/src/twcore/bots/hockeybot/hockeymediator/HockeyMediator.java
@@ -17,7 +17,7 @@ public interface HockeyMediator {
     */
     
     public void doReadyTeam(String name, String message);
-    public void notifyTime(short mins, short secs);
+    public void notifyTime(long mins, long secs);
     public void startPractice(String name, String squadAccepted);
